Correcciones

+++ src/naves/Nave.java
@@ -10,7 +10,7 @@ public class Nave {
 	protected Direccion direccion;
 	protected EstadoDeSalud estado;
 	protected List<SeccionDeNave> secciones = new LinkedList<SeccionDeNave>();
-	protected float porcentajeDeVida;
+	protected int porcentajeDeVida;
 
 	public Nave(int largoNave, Direccion direccionNave) {
 		this.largo = largoNave;
@@ -59,7 +59,7 @@ public class Nave {
 		}
 		else{
 			this.estado = EstadoDeSalud.DANADO;
-			this.porcentajeDeVida= ((seccionesDanadas)/2 + seccionesSanas) * (100/this.largo);			
+			this.porcentajeDeVida = (int)( ((seccionesDanadas)/2 + seccionesSanas) * ((float)100/this.largo));			
 		}
